+++ Network/API/Telegram/Bot/Object/Update/Message.hs
@@ -122,3 +122,25 @@ instance Persistable (Reply obj) => Persistable (Silently Reply obj) where
 	type Payload (Silently Reply obj) = Silently Reply (Payload (Reply obj))
 	payload (Silently x) = payload x <> singleton "disable_notification" (toJSON True)
 	endpoint (Silently x) = endpoint x
+
+data Edit b = Edit Int64 Int b
+
+instance Persistable (Edit Text) where
+	type Payload (Edit Text) = Edit Text
+	payload (Edit chat_id message_id text) = singleton "chat_id" (toJSON chat_id)
+		<> singleton "message_id" (toJSON message_id) <> singleton "text" (toJSON text)
+	endpoint _ = "editMessageText"
+
+instance Persistable (Edit Keyboard) where
+	type Payload (Edit Keyboard) = Edit Keyboard
+	payload (Edit chat_id message_id reply_markup) = singleton "chat_id" (toJSON chat_id)
+		<> singleton "message_id" (toJSON message_id) <> singleton "reply_markup" (toJSON reply_markup)
+	endpoint _ = "editMessageText"
+
+data Delete a = Delete Int64 Int
+
+instance Persistable (Delete Message) where
+	type Payload (Delete Message) = Delete Message
+	payload (Delete chat_id message_id) = singleton "chat_id" (toJSON chat_id)
+		<> singleton "message_id" (toJSON message_id)
+	endpoint _ = "deleteMessage"